+++ src/compile/compile.ts
@@ -5,33 +5,27 @@ import msgChannel, { OutputLevel } from "../utils/msg-channel.js";
 import statusBar from "../utils/status-bar.js";
 import { isFileModified } from "../utils/helper.js";
 import transform from "./transform.js";
+import { userConfig } from "../utils/config.js";
 
-const userConfig = {
-	outDir: "out",
-	rootDir: "src",
-};
-
-export const tscConfig = { compilerOptions: { module: 1, target: 8 } };
-
-export const workspaceFolder = vscode.workspace.workspaceFolders[0].uri.path;
-
-async function walkDir(source: string) {
-	const dirents: Dirent[] = await readdir(source, { withFileTypes: true });
+async function walkDir(workspaceFolder, source: string) {
+	const dirents: Dirent[] = await readdir(workspaceFolder + source, { withFileTypes: true });
 
 	const promises = [];
 	for (const dirent of dirents) {
 		const dirPath = `${source}/${dirent.name}`;
-		if (dirent.isDirectory()) walkDir(dirPath);
-		else if (isFileModified(dirPath)) promises.push(transform(dirPath));
+		if (dirent.isDirectory()) walkDir(workspaceFolder, dirPath);
+		else if (isFileModified(workspaceFolder + dirPath)) promises.push(transform(workspaceFolder, dirPath));
 	}
 
-	await Promise.all(promises);
+	return await Promise.all(promises);
 }
 
 export async function compileDir() {
 	statusBar.processing();
-	const rootDirPath = `${workspaceFolder}/${userConfig.rootDir}`;
-	await walkDir(rootDirPath).catch((err) => msgChannel.info(err, OutputLevel.Error));
+	const workspaceFolder = vscode.workspace.workspaceFolders[0].uri.path;
+	for (const rootDir of userConfig.rootDir) {
+		await walkDir(workspaceFolder, "/" + rootDir).catch((err) => msgChannel.info(err, OutputLevel.Error));
+	}
 	statusBar.done();
 }
 
@@ -41,7 +35,8 @@ export async function compileOnSave(textDocument: vscode.TextDocument): Promise<
 
 	statusBar.processing();
 	msgChannel.info(fileName);
-
-	await transform(textDocument.fileName).catch((err) => msgChannel.info(err, OutputLevel.Error));
+	const workspaceFolder = vscode.workspace.workspaceFolders[0].uri.path;
+	const filePath = textDocument.fileName.slice(workspaceFolder.length);
+	await transform(workspaceFolder, filePath).catch((err) => msgChannel.info(err, OutputLevel.Error));
 	statusBar.done();
 }