+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Change Log
 
 - Initial release
+
+- switch from swc to ts.transpileModule